requirements.txt
asgiref==3.7.2
Django==5.0.1
django-appconf==1.0.6
django-cleanup==8.0.0
django-extensions==3.2.3
django-filter==23.5
django-hashid-field==3.4.0
django-imagekit==5.0.0
django-leaflet==0.29.0
django-wkhtmltopdf==3.4.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
hashids==1.3.1
Markdown==3.5.2
pilkit==3.0
pillow==10.2.0
PyJWT==2.8.0
python-dotenv==1.0.0
pytz==2023.3.post1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2023.4
